refactor(webinst): Remove <changelog> node in info.xml (refs #4370)
The use of the <changelog/> node is deprecated, and replaced by a "changelog"
property (on the <module/> node) which contains a URL to the changelog.

The changelog, if it exists, must be specified in the module's "changelog"
property.

E.g. <module name="xxx" changelog="http://example.net/module/xxx/changelog/v1.2.3" />