fix(ihm): Fix editenumitemswidget.js UI (refs #4716)
Fix editenumitemswidget.js to add new enums to an ampty enum attrribute + fix
various misbehaviours (deletion of items, etc.).

The deletion of items has been changed to work on a new data-key property
instead of the data-index properties which ended up being out of sync when an
item was deleted (which prevented deletion of further items).