fix(CheckEnd): Store attributes during import (refs #5039)
Some imported parameters were not properly checked when updating an
existing family as the old family class was loaded in memory and would
not be updated with the newly defined parameters.

This commit modifies ImportDescription to store attributes defined, or
updated, in the current import session for checking before the
corresponding FDLGEN class gets regenerated and reloaded.