feat(DSearch): Set a default phpfunc for docid attributes in DSearch UI (refs #5447)
When composing a detailed search, or report, a docid with a custom
phpfunc using multiple attributes on input would not be functional and
usable for composing the criteria.

Thus, resetting the docid's phpfunc to a default phpfunc allows those
attributes to be usable when editing the criteria.

However, the side-effect is that legitimate custom phpfuncs won't be
used when composing the search's criteria.