feat(htmltext): Normalize HTML to prevent DOM document corruption (refs #5695)
Malformed HTML values could corrupt a document's DOM representation,
thus rendering the view or edition impossible.

To mitigate this situations, a "normalization" is performed on the HTML
values to ensure a malformed value cannot be entered or that a "safe"
form is used instead.

The "normalization" consists in loading the fragment with libxml's HTML
parser, and serialize it back to HTML. Warnings/errors during the
processing of the HTML fragment are reported alongside the "corrected"
HTML fragment.

- Doc::setValue()

  When the value is a malformed HTML, an error is returned indicating
  the reason (e.g. "error parsing attribute name", "Couldn't find end of
  Start Tag table", etc.) and the value is not set.

- DocHtmlFormat::formatHtmltext()

  When the value is a malformed HTML, a warning is reported through
  addWarningMsg() indicating the reason, and the normalized/corrected
  HTML is used.

- DocOooFormat::formatHtmltext()

  When the value is a malformed HTML, a warning is reported through
  addWarningMsg() indicating the reason, and the normalized/corrected
  HTML is used.