fix(wstart): Process all `Images` subdirs (refs #5431)
Initially, only images from `Images` dirs contained in a subdir of an
application (i.e. containing a `*.app` file) would be processed.

This was problematic as the content of the `FUSERS` dir is bundled by
two distinct modules: `dynacase-core` which bundle FUSERS's images in
`FUSERS/Images/` but not the `FUSERS/FUSERS.app` file, and
dynacase-admin-fusers-ui which bundles the remaining
`FUSERS/FUSERS.app` file. This means that the `FUSERS/Images` images
would not be processed if only `dynacase-core` was installed, and they
would be process when `dynacase-admin-fusers-ui` was installed.

So, this commit disables the filter on the presence of a `*.app` file in
order to process all first-level dirs having a `Images` subdir (except
in the `supervisor` first-level dir).