feat(CORE): Remove hardcoded references to the context's root dir (refs #5939)
Previously, the context's root dir was hardcoded in the `CORE_PUBDIR`
paramv in database and the `$pubdir` variable in Lib.Prefix.php(.in).

Now, the context's root dir is dynamically computed from the
Lib.Prefix.php pathname and available through the `$pubdir` variable and
the DEFAULT_PUBDIR constant, and the `CORE_PUBDIR` paramv has been
removed.