feat(processExecute): Use sub-processes to execute doctimers and EXECs documents (refs #6223)
All the doctimers and EXECs documents are processed in a separate
sub-process which allows to keep processing the remaining timers and
documents whenever one fails with a PHP Fatal error or an uncaught
exception.

The failures of the sub-processes are then handled with the new WSH
error message notification mechanism introduced by #6535.